+++ modules/farm_rothamsted_experiment/src/Form/ExperimentBoundaryForm.php
@@ -84,16 +84,7 @@ class ExperimentBoundaryForm extends ExperimentFormBase {
     $form['required_message'] = [
       '#type'  => 'html_tag',
       '#tag'   => 'p',
-      '#value' => $this->t('The experiment code and experiment locations are required to create an experiment boundary. Verify that these are correct before creating the experiment boundary.'),
-    ];
-
-    // Study period.
-    $form['study_period_id'] = [
-      '#type' => 'textfield',
-      '#title' => $this->t('Study Period ID'),
-      '#description' => $this->t('The unique identifier for the study, for example 2020/R/CS/790.'),
-      '#default_value' => $plan->get('study_period_id')->value,
-      '#required' => TRUE,
+      '#value' => $this->t('The experiment location is required to create an experiment boundary. Verify that this is correct before creating the experiment boundary.'),
     ];
 
     // Location for the experiment boundary parents.
@@ -156,9 +147,8 @@ class ExperimentBoundaryForm extends ExperimentFormBase {
     // Get the plan.
     $plan = Plan::load($form_state->getValue('plan_id'));
 
-    // Set the study_period_id.
-    $code = $form_state->getValue('study_period_id');
-    $plan->set('study_period_id', $code);
+    // Get the study_period_id.
+    $study_period = $plan->get('study_period_id')->getValue();
 
     // Set the experiment location.
     $location = $form_state->getValue('location');
@@ -168,7 +158,7 @@ class ExperimentBoundaryForm extends ExperimentFormBase {
     $boundary = Asset::create([
       'type' => 'land',
       'land_type' => 'experiment_boundary',
-      'name' => $this->t('@code (@plan_name): Experiment Boundary', ['@code' => $code, '@plan_name' => $plan->label()]),
+      'name' => $this->t('@study_period (@plan_name): Experiment Boundary', ['@study_period' => $study_period, '@plan_name' => $plan->label()]),
       'status' => 'active',
       'parent' => $location,
       'is_fixed' => TRUE,