+++ modules/farm_rothamsted_experiment/src/Plugin/Plan/PlanType/Experiment.php
@@ -48,6 +48,7 @@ class Experiment extends FarmPlanType {
         'type' => 'string',
         'label' => $this->t('Study Period ID'),
         'description' => $this->t('The unique identifier for the study, for example 2020/R/CS/790.'),
+        'required' => TRUE,
       ],
       'cost_code' => [
         'type' => 'string',